requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
openai==1.12.0
pyyaml==6.0.1
python-dotenv==1.0.1
flask==3.0.2
ratelimit==2.2.1
lxml==5.1.0
